move `stories/` to `tests/` & fix tests
