+++ src/compiler/transform/remove-export-default.test.ts
@@ -16,7 +16,7 @@ describe(removeExportDefault.name, () => {
   it('removes pre-transformed export default correctly', async ({ expect }) => {
     const compiledPreTransformCode = fs
       .readFileSync(
-        path.resolve(__dirname, '../../../test/__compiled__/pre-transform/Example.stories.dev.js')
+        path.resolve(__dirname, '../../../tests/__compiled__/pre-transform/Example.stories.dev.js')
       )
       .toString();
     const compiledASTNodes = await extractCompiledASTNodes({
