+++ src/parser/extract/svelte/module-nodes.test.ts
@@ -32,15 +32,7 @@ describe(extractModuleNodes.name, () => {
     expect(extractModuleNodes({ module })).rejects.toThrowErrorMatchingInlineSnapshot(`
       [SB_SVELTE_CSF_PARSER_EXTRACT_SVELTE_0003 (MissingDefineMetaImportError): Stories file: <path not specified>
       doesn't have a 'defineMeta' imported from the "@storybook/addon-svelte-csf" package inside the module tag.
-      ('<script context="module"> <!-- ... --> </script>').
-
-      Make sure this stories file has initial code snippet in order for this addon to work correctly:
-
-      <script context="module">
-      import { defineMeta } from "@storybook/addon-svelte-csf";
-
-      const { Story } = defineMeta({});
-      </script>]
+      ('<script context="module"> <!-- ... --> </script>').]
     `);
   });
 
@@ -55,15 +47,7 @@ describe(extractModuleNodes.name, () => {
 
     expect(extractModuleNodes({ module })).rejects.toThrowErrorMatchingInlineSnapshot(`
       [SB_SVELTE_CSF_PARSER_EXTRACT_SVELTE_0004 (MissingDefineMetaVariableDeclarationError): Stories file: <path not specified>
-      doesn't have 'defineMeta' call used for variable declaration inside the module tag ('<script context="module"> <!-- ... --> </script>').
-
-      Make sure this stories file has initial code snippet in order for this addon to work correctly:
-
-      <script context="module">
-      import { defineMeta } from "@storybook/addon-svelte-csf";
-
-      const { Story } = defineMeta({});
-      </script>]
+      doesn't have 'defineMeta' call used for variable declaration inside the module tag ('<script context="module"> <!-- ... --> </script>').]
     `);
   });
 
@@ -78,15 +62,7 @@ describe(extractModuleNodes.name, () => {
 
     expect(extractModuleNodes({ module })).rejects.toThrowErrorMatchingInlineSnapshot(`
       [SB_SVELTE_CSF_PARSER_EXTRACT_SVELTE_0004 (MissingDefineMetaVariableDeclarationError): Stories file: <path not specified>
-      doesn't have 'defineMeta' call used for variable declaration inside the module tag ('<script context="module"> <!-- ... --> </script>').
-
-      Make sure this stories file has initial code snippet in order for this addon to work correctly:
-
-      <script context="module">
-      import { defineMeta } from "@storybook/addon-svelte-csf";
-
-      const { Story } = defineMeta({});
-      </script>]
+      doesn't have 'defineMeta' call used for variable declaration inside the module tag ('<script context="module"> <!-- ... --> </script>').]
     `);
   });
