+++ src/runtime/emit-code.ts
@@ -102,7 +102,7 @@ const getFunctionName = (fn: Function & { getMockName?: () => string }) => {
  * convert a value to a stringified version
  */
 const valueToString = (value: any): string => {
-  if (value[Symbol.for('svelte.snippet')]) {
+  if (typeof value === 'object' && value[Symbol.for('svelte.snippet')]) {
     return 'snippet';
   }
