+++ tests/stories/BackgroundDecorator.svelte
