+++ tsconfig.json
@@ -24,5 +24,5 @@
       }
     ]
   },
-  "include": ["src/**/*", "examples/**/*"]
+  "include": ["src/**/*", "examples/**/*", "tests/**/*"]
 }
